backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
httpx
google-cloud-texttospeech
